backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-JWT-Extended==4.6.0
flask-socketio==5.3.6
supabase==2.13.0
python-dotenv==1.0.1
requests==2.31.0
geopy==2.4.1
gunicorn==22.0.0
eventlet==0.36.1
razorpay==1.4.2
openai>=1.0.0
anthropic>=0.18.0
reportlab==4.0.7
pyjwt==2.8.0
websocket-client==1.7.0
pyotp==2.9.0
apscheduler==3.10.4
python-slugify==8.0.4
markdown==3.5.1
qrcode==7.4.2
pillow>=10.3.0
httpx==0.28.1
slack-sdk==3.25.0
websockets>=13.0
jinja2>=3.1.2
